==> google_trend/03_google_trend_dynamic_factor_analysis_by_rstan.R <==
# setwd("//media//kswada//MyFiles//R//google_trend//")
setwd("//media//kswada//MyFiles//R//Bayesian_inference//google_trend//")

packages <- c("dplyr", "rstan")
purrr::walk(packages, library, character.only = TRUE, warn.conflicts = FALSE)



# ------------------------------------------------------------------------------
# data:  google trend
# ------------------------------------------------------------------------------

library(tidyverse)

s.dfData2 <- read_csv("dfData2.csv", locale = locale(encoding = "cp932"))


str(s.dfData2)


dim(s.dfData2)


car::some(s.dfData2)



# ------------------------------------------------------------------------------
# Dynamic Fdactor Analysis:  functions
# ------------------------------------------------------------------------------

makeDFA108.1 <- function(lDataDFA100, STANDIR){

  # purpose: 8因子DFAモデルの推定
  # args: 
  #   lDataDFA100: generated by makeDataDFA100.1()
  #   STANDIR:     Stanファイルの所在
  # return: 
  #   sub_Estimate_Stan.409()の返し値
  
  lModel <- sub_Estimate_Stan.409 (
    mgData     = lDataDFA100$mgData,
    mgCov      = lDataDFA100$mgCov,
    nNumFactor = 8,
    nChain     = 2,
    nIter      = 3000,
    nWarm      = 1000, 
    sStanDir   = STANDIR
  )
  return(lModel)
}


# ----------
sub_Estimate_Stan.409 <- function(
  mgData      = mgData,
  mgCov       = mgCov,
  nNumFactor  = 2,
  nIter       = 400,
  nWarm       = 200,
  nChain      = 1, 
  sStanDir    = "."
) {
  
  # - - - - 
  sModel <- paste0(STANDIR, "/sub_Estimate_Stan.409.stan")
  # - - - - 
  
  stopifnot(nrow(mgData) == nrow(mgCov))
  
  # mgY: データ(時点が列)
  mgY <- t(mgData)
  
  # mgC: 共変量(時点が列)
  mgC <- t(mgCov)

  # nNumTime: 時点数
  nNumTime <- ncol(mgY)
  
  # nNumSeries: 時系列の数
  nNumSeries <- nrow(mgY)
  
  # nNumFactor: 因子数
  nNumFactor <- nNumFactor 
  
  # nNumC:共変量の数
  nNumC <- nrow(mgC)
  
  # ----------
  # 標準化
  for (i in 1:nNumSeries) {
    mgY[i, ] = scale(mgY[i, ], center = TRUE, scale = TRUE)
  }
  
  # ----------
  # mbZType: 因子負荷行列と同じサイズの行列
  #          0: 0に固定する場所(因子fの上から(f-1)行)
  #          1: 自由推定する場所
  mbZType <- matrix(1, nNumSeries, nNumFactor)
  for (f in 2:nNumFactor){
    mbZType[1:(f-1), f] <- 0
  }
  
  # ----------
  # mnZIndex: 因子負荷行列と同じサイズの行列
  #           0: 0に固定する場所
  #           1からはじまる連番: 自由推定する場所
  mnZIndex <- matrix(cumsum(mbZType), nrow=nNumSeries) * mbZType
  
  # ----------
  # dfZIndex: 行は因子負荷行列の要素
  dfZIndex <- crossing(nRow = 1:nrow(mbZType), nCol = 1:ncol(mbZType))
  
  anType  <- mbZType[as.matrix(dfZIndex)]
  
  dfZIndex <- dfZIndex %>%
    mutate(nType = anType) %>%
    arrange(nType, nCol, nRow) %>%
    group_by(nType) %>%
    mutate(nID = seq_along(nType)) %>%
    ungroup() 
  
  mnZLoc0 <- dfZIndex %>% filter(nType == 0) %>% dplyr::select(nRow, nCol) %>% as.matrix(.)
  mnZLoc1 <- dfZIndex %>% filter(nType == 1) %>% dplyr::select(nRow, nCol) %>% as.matrix(.)
  nNumZ0 <- nrow(mnZLoc0)
  nNumZ1 <- nrow(mnZLoc1)
  
  
  # ----------
  lData <- list(
    nNumTime,
    nNumSeries,
    nNumFactor,
    nNumZ0,
    nNumZ1,
    mnZLoc0, 
    mnZLoc1,
    mgY,
    nNumC,
    mgC
  )
  
  # ----------
  asPars <- c("mgX", "mgZ", "gSigma", "mgBeta", "mgPred", "log_lik")
  
  # ----------
  lStan <- rstan::stan(
    data    = lData,
    file    = sModel,
    pars    = asPars,
    chains  = nChain,
    iter    = nIter,
    warmup  = nWarm,
    control = list(adapt_delta = 0.99, max_treedepth = 15),
    refresh = 50,
    verbose = TRUE,
    seed    = 123
  )
  return(lStan)
}


# ------------------------------------------------------------------------------
# Dynamic Fdactor Analysis:  MCMC estimation
# ------------------------------------------------------------------------------

STANDIR     <- "./stan"

s.lDFA108 <- makeDFA108.1(s.lDataDFA100, STANDIR)
